refactor: improve SinglePlaylist and SinglePlaylistTitle components; rename and delete functions are together + confirmation modal
